Removed duplicated variable declaration (internal_dnn_convolve_winograd_st_bwd_nhwc_custom_alpha6). Removed unused variable (Owp). Convert loop upper bound to integer in order to avoid warning about signed/unsigned comparison.

+++ src/template/libxsmm_dnn_convolution_winograd_backward_custom_custom_input_trans_alpha6_avx512.tpl.c
@@ -41,8 +41,8 @@ const int t_pad = (handle->desc.H - handle->ofh)/2 + 1;
 __m512 T[ALPHA][ALPHA]; /* FIXME: too big and causing spills */
 __m512 t0, t1, t2, t3, t4, t5;
 
-for (tj = 0; tj < handle->cwino_bwd.jtiles; tj++) {
-  for (ti = 0; ti < handle->cwino_bwd.itiles; ti++) { /* for each tile */
+for (tj = 0; tj < (int)handle->cwino_bwd.jtiles; tj++) {
+  for (ti = 0; ti < (int)handle->cwino_bwd.itiles; ti++) { /* for each tile */
     if (ti*((ALPHA)-2) >= l_pad && ti*((ALPHA)-2) + (ALPHA) <= (handle->ofw + l_pad) &&
         tj*((ALPHA)-2) >= t_pad && tj*((ALPHA)-2) + (ALPHA) <= (handle->ofh + t_pad)) { /* common case */
